refactor: 시간 관련 계산 함수 utils로 이동
